XC-372: move tests
